Implement IdeaItem serialization and deserialization
This commit also adds the use of the serialization and deserialization
items to the application, so that now when items are added they will be
saved on close and reloaded when the application is opened again. Items
are also now displayed in the reviewPane and able to be demo-reviewed in
the beta interface.